refactor logger
