+++ core/pkg/server/config.go
@@ -75,7 +75,7 @@ func (runConfig *RunConfig) ApplyChangeRecord(
 		); err != nil {
 			onError(
 				fmt.Errorf(
-					"Failed to unmarshall JSON for config key %v: %w",
+					"failed to unmarshall JSON for config key %v: %w",
 					path,
 					err,
 				),
@@ -162,7 +162,7 @@ func (runConfig *RunConfig) Serialize(format ConfigFormat) ([]byte, error) {
 		return json.Marshal(valueConfig)
 	}
 
-	return nil, fmt.Errorf("Unknown format: %v", format)
+	return nil, fmt.Errorf("unknown format: %v", format)
 }
 
 // Uses the given subtree for keys that aren't already set.
@@ -280,7 +280,7 @@ func getOrMakeSubtree(
 		subtree, ok := node.(RunConfigDict)
 		if !ok {
 			return nil, fmt.Errorf(
-				"Config value at path %v is type %T, not a map",
+				"config value at path %v is type %T, not a map",
 				path,
 				node,
 			)
