+++ core/pkg/server/writer.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"context"
+	"log/slog"
 	"os"
 	"sync"
 
@@ -57,7 +58,7 @@ type Writer struct {
 func NewWriter(ctx context.Context, logger *observability.CoreLogger, opts ...WriterOption) *Writer {
 	w := &Writer{
 		ctx:    ctx,
-		logger: logger,
+		logger: observability.NewCoreLogger(slog.New(logger.Handler().WithGroup("writer"))),
 		wg:     sync.WaitGroup{},
 	}
 	for _, opt := range opts {
@@ -78,19 +79,22 @@ func (w *Writer) startStore() {
 	w.store = NewStore(w.ctx, w.settings.GetSyncFile().GetValue(), w.logger)
 	err = w.store.Open(os.O_WRONLY)
 	if err != nil {
-		w.logger.CaptureFatalAndPanic("writer: error creating store", err)
+		w.logger.Error("failed to open store", "error", err)
+		panic(err)
+		// w.logger.CaptureFatalAndPanic("writer: error creating store", err)
 	}
 
 	w.wg.Add(1)
 	go func() {
 		for record := range w.storeChan {
 			if err = w.store.Write(record); err != nil {
-				w.logger.Error("writer: error storing record", "error", err)
+				w.logger.Error("failed to write to store", "error", err)
 			}
 		}
 
 		if err = w.store.Close(); err != nil {
-			w.logger.CaptureError("writer: error closing store", err)
+			w.logger.Error("failed to close store", "error", err)
+			// w.logger.CaptureError("writer: error closing store", err)
 		}
 		w.wg.Done()
 	}()
@@ -99,11 +103,12 @@ func (w *Writer) startStore() {
 // do is the main loop of the writer to process incoming messages
 func (w *Writer) Do(inChan <-chan *service.Record) {
 	defer w.logger.Reraise()
-	w.logger.Info("writer: started", "stream_id", w.settings.RunId)
+	w.logger.Info("started writer")
 
 	w.startStore()
 
 	for record := range inChan {
+		w.logger.Debug("received record", "record", record.RecordType)
 		w.handleRecord(record)
 	}
 	w.Close()
@@ -117,7 +122,7 @@ func (w *Writer) Close() {
 	if w.storeChan != nil {
 		close(w.storeChan)
 	}
-	w.logger.Info("writer: closed", "stream_id", w.settings.RunId)
+	w.logger.Info("closed writer")
 }
 
 // handleRecord Writing messages to the append-only log,
@@ -125,7 +130,6 @@ func (w *Writer) Close() {
 // We ensure that the messages are written to the log
 // before they are sent to the server.
 func (w *Writer) handleRecord(record *service.Record) {
-	w.logger.Debug("write: got a message", "record", record, "stream_id", w.settings.RunId)
 	switch record.RecordType.(type) {
 	case *service.Record_Request:
 		w.sendRecord(record)
